adding pdft_feff stuff, nothing new yet, just some interface stuff
